# api-gateway/api-gateway-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/create": {
            "post": {
                "description": "Register a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register a user",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateUserResponse"
                        }
                    }
                }
            }
        },
...
        "/user/list": {
            "get": {
                "description": "retrieve the list of users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get the list of users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListUsersResponse"
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "login a user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "login a user",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.LoginUserResponse"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "Access-Token"
                            },
                            "X-Refresh-Token": {
                                "type": "string",
                                "description": "Refresh-Token"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.GetUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListUsersResponse": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.GetUserResponse"
                    }
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.LoginUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.LoginUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}